登录
新增了一些功能

+++ test/vue/src/views/Employee.vue
@@ -10,14 +10,16 @@
 
     <div class="card" style="margin-bottom: 5px;">
         <el-button type="primary" @click="handleAdd()">新增</el-button>
-        <el-button type="primary">批量删除</el-button>
+        <el-button type="primary" @click="deleteBatch()">批量删除</el-button>
         <!-- <el-button type="primary">导入</el-button>
         <el-button type="primary">导出</el-button> -->
     </div>
 
     <div class="card" style="margin-bottom: 5px;">
         <!-- 用户数据表格 -->
-        <el-table :data="data.tableData">
+        <el-table :data="data.tableData" @selection-change="deleteBatchEmployeeById">
+            <el-table-column type="selection" width="55" />
+            <el-table-column label="账号" prop="username" />
             <el-table-column label="名称" prop="name" />
             <el-table-column label="性别" prop="sex" />
             <el-table-column label="工号" prop="no" />
@@ -29,7 +31,7 @@
                 <template #default="scope">
                     <el-button link type="primary" icon="edit" @click="handleUpdate(scope.row)"> 编辑
                     </el-button>
-                    <el-button link type="danger" icon="delete" @click="deleteEmployeeByNo(scope.row)"> 删除
+                    <el-button link type="danger" icon="delete" @click="deleteEmployeeById(scope.row)"> 删除
                     </el-button>
                 </template>
             </el-table-column>
@@ -44,21 +46,28 @@
         </div>
 
         <!-- 新增的弹窗 -->
-        <el-dialog title="新增员工" v-model="data.formVisible" width="500px">
-            <el-form :model="data.form" label-width="80px" style="padding-right: 40px; padding-top: 20px;">
+        <!-- destroy-on-close 在弹窗关闭时销毁掉这个弹窗 -->
+        <el-dialog title="新增员工" v-model="data.formVisible" width="500px" destroy-on-close>
+            <el-form ref="formRef" :rules="data.rules" :model="data.form" label-width="80px"
+                style="padding-right: 40px; padding-top: 20px;">
 
-                <el-form-item label="名称">
+                <el-form-item label="账号" prop="username">
+                    <el-input v-model="data.form.username" autocomplete="off" placeholder="请输入账号" />
+                </el-form-item>
+
+                <el-form-item label="名称" prop="name">
                     <el-input v-model="data.form.name" autocomplete="off" placeholder="请输入名称" />
                 </el-form-item>
 
                 <el-form-item label="性别">
                     <el-radio-group v-model="data.form.sex">
-                        <el-radio value="1" label="男"></el-radio>
-                        <el-radio value="0" label="女"></el-radio>
+                        <!-- 在编辑员工信息时 该值是动态的 所以要绑定 -->
+                        <el-radio :value="1" label="男"></el-radio>
+                        <el-radio :value="0" label="女"></el-radio>
                     </el-radio-group>
                 </el-form-item>
 
-                <el-form-item label="工号">
+                <el-form-item label="工号" prop="no">
                     <el-input v-model="data.form.no" autocomplete="off" placeholder="请输入工号" />
                 </el-form-item>
 
@@ -94,7 +103,7 @@
 
 <script setup>
 import request from '@/utils/request.js';
-import { ElMessage } from 'element-plus';
+import { ElMessage, ElMessageBox } from 'element-plus';
 import { ref, reactive, onMounted } from 'vue';
 import Api from '@/const/Api'
 
@@ -108,25 +117,44 @@ const data = reactive({
     formVisible: false,
     editVisible: false,
     form: {},
-    department: {}
+    department: {},
+    ids: [],
+    rules: {
+        username: [
+            // 是否必填 触发校验的提示词 触发时机 
+            { required: true, message: '请输入账号', trigger: 'blur' }
+        ],
+        name: [
+            // 是否必填 触发校验的提示词 触发时机 
+            { required: true, message: '请输入名称', trigger: 'blur' }
+        ],
+        no: [
+            // 是否必填 触发校验的提示词 触发时机 
+            { required: true, message: '请输入工号', trigger: 'blur' }
+        ]
+    }
 })
+
+const formRef = ref()
 //调用java中的
 // 将数据库的数据写入到表格中
 //表格查询功能
+// TODO
 async function load() {
     const res = await request.get(Api.selectPage, {
         params: {
+            name: data.name,
             pageNum: data.pageNum,
-            pageSize: data.pageSize,
-            name: data.name
+            pageSize: data.pageSize
         }
     })
     data.tableData = res.data.list
     data.total = res.data.total
+
 }
 
 //表格查询重置功能
-const reset = () => {
+function reset() {
     data.name = null
     load()
 }
@@ -143,7 +171,7 @@ function handleAdd() {
     data.form = {}
     data.formVisible = true
 }
-
+// 编辑员工数据的功能
 function handleUpdate(row) {
     getDepartment()
     // 将对象中的数据提取给data.form 而不是直接修改对象row 
@@ -153,7 +181,12 @@ function handleUpdate(row) {
 
 //表格新增员工弹窗的保存功能
 async function save() {
-    data.form.id ? edit() : add()
+    // 表单校验
+    formRef.value.validate((valid) => {
+        if (valid) {
+            data.form.id ? edit() : add()
+        }
+    })
 }
 //表格新增员工的新增功能
 async function add() {
@@ -181,14 +214,64 @@ async function edit() {
 }
 
 // 表格删除员工信息功能
-async function deleteEmployeeByNo(row) {
-    data.form = JSON.parse(JSON.stringify(row))
-    const res = await request.delete(Api.deleteEmployeeById, {
-        params: {
-            "id": data.form.id
+// action是按钮值 confirm是确认
+async function deleteEmployeeById(row) {
+    // 一个确认信息的弹窗 第一个参数是 提示信息 第二个参数是 提示标题 第三个参数是弹窗类型 有警告warning 成功succes 错误error 和 信息info 
+    ElMessageBox.confirm('删除后数据无法恢复,您确认删除吗?', '删除确认', {
+        type: 'warning',
+        // 弹窗关闭后的操作
+        callback: async (action) => {
+            //当点击确认时执行
+            if (action == 'confirm') {
+                data.form = JSON.parse(JSON.stringify(row))
+                const res = await request.delete(Api.deleteEmployeeById, {
+                    params: {
+                        id: data.form.id
+                    }
+                })
+                if (res.code === 200) {
+                    ElMessage.success('操作成功')
+                    load()
+                }
+            }
+            else {
+                ElMessage.error(res.msg)
+            }
         }
     })
-    load()
+}
+
+//批量删除
+function deleteBatchEmployeeById(rows) {
+    // 这个意思是将 rows中的每个对象提取成row 然后箭头后面将对象进行处理
+    data.ids = rows.map(row => row.id)
+}
+
+async function deleteBatch() {
+    if (data.ids.length === 0) {
+        ElMessage.warning('请选择数据')
+        return;
+    }
+    ElMessageBox.confirm('删除后数据无法恢复,您确认删除吗?', '删除确认', {
+        type: 'warning',
+        callback: async (action) => {
+            //当点击确认时执行
+            if (action == 'confirm') {
+                // 这里以因为是delet请求所以没有请求体 只能写成配置的样式 不能直接传对象
+                const res = await request.delete(Api.deleteBatch, {
+                    data: data.ids
+                })
+                if (res.code === 200) {
+                    ElMessage.success('操作成功')
+                    load()
+                }
+            }
+            else {
+                ElMessage.error(res.msg)
+            }
+        }
+    })
+
 }
 
 onMounted(() => {
